+++ packages/app-admin/src/components/FileManager/outputFileSelectionError.ts
@@ -26,4 +26,5 @@ export default errors => {
         case "multipleNotAllowed":
             return "Only one file allowed.";
     }
+    return null;
 };