+++ packages/app-headless-cms/src/admin/components/ContentEntryForm/useContentEntryForm.ts
@@ -40,6 +40,7 @@ export interface UseContentEntryFormParams {
     entry?: { [key: string]: any };
     onChange?: FormOnSubmit;
     onSubmit?: FormOnSubmit;
+    addEntryToListCache: boolean;
 }
 
 export function useContentEntryForm(params: UseContentEntryFormParams): UseContentEntryForm {
@@ -97,7 +98,14 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
                         return;
                     }
                     resetInvalidFieldValues();
-                    GQLCache.addEntryToListCache(contentModel, cache, entry, listQueryVariables);
+                    if (params.addEntryToListCache) {
+                        GQLCache.addEntryToListCache(
+                            contentModel,
+                            cache,
+                            entry,
+                            listQueryVariables
+                        );
+                    }
                 }
             });
             setLoading(false);
@@ -110,10 +118,14 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
             }
             resetInvalidFieldValues();
             showSnackbar(`${contentModel.name} entry created successfully!`);
-            goToRevision(entry.id);
+            if (typeof params.onSubmit === "function") {
+                params.onSubmit(entry);
+            } else {
+                goToRevision(entry.id);
+            }
             return entry;
         },
-        [contentModel.modelId, listQueryVariables]
+        [contentModel.modelId, listQueryVariables, params.onSubmit, params.addEntryToListCache]
     );
 
     const updateContent = useCallback(
@@ -161,12 +173,7 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
                     GQLCache.addRevisionToRevisionsCache(contentModel, cache, newRevision);
 
                     showSnackbar("A new revision was created!");
-
-                    history.push(
-                        `/cms/content-entries/${contentModel.modelId}?id=${encodeURIComponent(
-                            newRevision.id
-                        )}`
-                    );
+                    goToRevision(newRevision.id);
                 }
             });
             setLoading(false);