+++ packages/app-headless-cms/src/admin/plugins/fieldRenderers/ref/components/ContentEntriesMultiAutoComplete.tsx
@@ -1,10 +1,12 @@
-import React from "react";
+import React, { useCallback } from "react";
 import debounce from "lodash/debounce";
 import { MultiAutoComplete } from "@webiny/ui/AutoComplete";
 import { Link } from "@webiny/react-router";
 import { i18n } from "@webiny/app/i18n";
 import { useReferences } from "./useReferences";
 import { renderItem } from "./renderItem";
+import NewRefEntryFormDialog, { NewEntryButton } from "./NewRefEntryFormDialog";
+import { useNewRefEntry } from "../hooks/useNewRefEntry";
 
 const t = i18n.ns("app-headless-cms/admin/fields/ref");
 
@@ -12,6 +14,8 @@ const warn = t`Before publishing the main content entry, make sure you publish t
 
 function ContentEntriesMultiAutocomplete({ bind, field }) {
     const { options, setSearch, entries, loading, onChange } = useReferences({ bind, field });
+
+    const { renderNewEntryModal, refModelId, helpText } = useNewRefEntry({ field });
 
     const entryWarning = ({ id, modelId, name, published }, index) =>
         !published && (
@@ -28,6 +32,42 @@ function ContentEntriesMultiAutocomplete({ bind, field }) {
         warning = warn({
             entries: <>{warning.map(entryWarning)}</>
         });
+    }
+
+    const refEntryOnChange = useCallback(
+        value => {
+            /**
+             * Append new selected entry at the end of existing entries.
+             */
+            onChange([...entries, value]);
+        },
+        [onChange, entries]
+    );
+
+    if (renderNewEntryModal) {
+        return (
+            <NewRefEntryFormDialog modelId={refModelId} onChange={refEntryOnChange}>
+                <MultiAutoComplete
+                    {...bind}
+                    renderItem={renderItem}
+                    renderListItemLabel={renderItem}
+                    useMultipleSelectionList
+                    onChange={onChange}
+                    loading={loading}
+                    value={entries}
+                    options={options}
+                    label={field.label}
+                    onInput={debounce(setSearch, 250)}
+                    description={
+                        <>
+                            {field.helpText}
+                            {warning}
+                        </>
+                    }
+                    noResultFound={<NewEntryButton />}
+                />
+            </NewRefEntryFormDialog>
+        );
     }
 
     return (
@@ -48,6 +88,7 @@ function ContentEntriesMultiAutocomplete({ bind, field }) {
                     {warning}
                 </>
             }
+            noResultFound={helpText}
         />
     );
 }