+++ packages/app-security/src/contexts/Security.tsx
@@ -2,13 +2,13 @@ import minimatch from "minimatch";
 import React, { useState, useMemo, Dispatch, SetStateAction, useCallback } from "react";
 import { SecurityIdentity, SecurityPermission } from "~/types";
 
-export const SecurityContext = React.createContext<SecurityContextValue>(null);
-
-export interface SecurityContextValue {
+export interface SecurityContext {
     identity: SecurityIdentity | null;
     setIdentity: Dispatch<SetStateAction<SecurityIdentity>>;
     getPermission<T extends SecurityPermission = SecurityPermission>(name: string): T;
 }
+
+export const SecurityContext = React.createContext<SecurityContext>(null);
 
 export const SecurityProvider = props => {
     const [identity, setIdentity] = useState<SecurityIdentity>(null);